app/forum/requirements-dev.txt
-r requirements.txt  # includes the base.txt requirements file

coverage
django-debug-toolbar
hacking
